- Extracted BuildSpecCmd with impl
