+++ client/node-common/src/lib.rs
@@ -17,3 +17,5 @@
 pub mod service;
 
 pub mod command;
+
+pub mod cli;
